refactor: consolidate color formatting into core
The colorjs space-alias map + Color construction + per-format rendering
were duplicated across blocks and mcp (the source of the #1108/#1123
drift). Extract them to core:

- core/color-formats.ts: ColorFormat type + COLOR_FORMATS list, colorjs-free
  so the addon manager bundle can import the list (it previously kept a
  local copy because blocks' barrel pulls virtual:swatchbook/tokens).
- core/format-color.ts: the rendering kernel (space-alias map, Color
  construction, hex/rgb/hsl/oklch + gamut), moved verbatim from blocks.
- blocks/format-color.ts: re-exports core, output unchanged.
- mcp/format-color.ts: thin wrapper that keeps its format field, full-JSON
  raw, and null-on-invalid contract, delegating rendering to core. MCP
  color output now uses the canonical rendering (integer rgb, 6-digit hex)
  instead of colorjs toString() percentages/short hex.
- addon: imports the type + list from core's colorjs-free leaf.

Issue #1116 said three copies (blocks/mcp/addon); the addon only ever had
the format-name list, so it was two formatters + a triplicated type/list.

Closes #1116
Refs #1117
