+++ .changeset/color-format-core.md
@@ -1,0 +1,6 @@
+---
+'@unpunnyfuns/swatchbook-core': patch
+'@unpunnyfuns/swatchbook-mcp': patch
+---
+
+consolidate color formatting into core (new color-formats and format-color modules shared by blocks/mcp/addon); MCP color output now uses the shared canonical rendering (integer rgb, 6-digit hex)
